Fix token view recycling issue when scrolling
By design, the RecyclerView used to show the token list can reuse a
previous ViewHolder instance. If a code was displayed in that instance,
it appears in the recycled view if this part is not clearly reset when
bound to a new token.

To fix the issue, add this needed reset checking the code state in
bind().